+++ rasa_core/server.py
@@ -23,10 +23,12 @@ def create_argument_parser():
             description='starts server to serve an agent')
     parser.add_argument(
             '-d', '--core',
+            required=True,
             type=str,
             help="core model to run with the server")
     parser.add_argument(
             '-u', '--nlu',
+            required=True,
             type=str,
             help="nlu model to run with the server")
     parser.add_argument(
@@ -36,6 +38,7 @@ def create_argument_parser():
     parser.add_argument(
             '-v', '--verbose',
             default=True,
+            action="store_true",
             help="use verbose logging")
     parser.add_argument(
             '-o', '--log_file',