+++ CHANGELOG.rst
@@ -20,6 +20,7 @@ Removed
 
 Fixed
 -----
+- fixed http error if action listen is not confirmed. #42
 
 [0.7.1] - 2017-10-06
 ^^^^^^^^^^^^^^^^^^^^