+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Added
 -----
 - added support for conditional checkpoints. a checkpoint can be restricted to
   only allow one to use it if certain slots are set. see docs for details
+- utterance templates in domain yaml support buttons and images
 
 Changed
 -------