+++ CHANGELOG.rst
@@ -14,12 +14,14 @@ Added
 
 Changed
 -------
+- small dropout in standard keras model to decrease reliance on exact intents
 
 Removed
 -------
 
 Fixed
 -----
+- fixed http error if action listen is not confirmed. #42
 
 [0.7.1] - 2017-10-06
 ^^^^^^^^^^^^^^^^^^^^