+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Added
 - added support for conditional checkpoints. a checkpoint can be restricted to
   only allow one to use it if certain slots are set. see docs for details
 - utterance templates in domain yaml support buttons and images
+- validate domain yaml and raise exception on invalid file
 
 Changed
 -------