+++ CHANGELOG.rst
@@ -8,6 +8,11 @@ This project adheres to `Semantic Versioning`_ starting with version 0.7.0.
 ^^^^^^^^^^^^^^^^^^^^^^^^
 
 .. note:: This version is not yet released and is under active development.
+
+Added
+-----
+- support for new dataflow data format (previously api.ai)
+
 Fixed
 -----
 - Fixed loading of default model from S3. Fixes #633