+++ CHANGELOG.rst
@@ -11,15 +11,23 @@ This project adheres to `Semantic Versioning`_ starting with version 0.2.0.
 
 Added
 -----
+- added facebook bot example
+- added support for conditional checkpoints. a checkpoint can be restricted to
+  only allow one to use it if certain slots are set. see docs for details
+- utterance templates in domain yaml support buttons and images
+- validate domain yaml and raise exception on invalid file
+- ``run`` script to load models and handle messages from an input channel
 
 Changed
 -------
+- small dropout in standard keras model to decrease reliance on exact intents
 
 Removed
 -------
 
 Fixed
 -----
+- fixed http error if action listen is not confirmed. #42
 
 [0.7.1] - 2017-10-06
 ^^^^^^^^^^^^^^^^^^^^