+++ CHANGELOG.rst
@@ -12,12 +12,13 @@ This project adheres to `Semantic Versioning`_ starting with version 0.7.0.
 Added
 -----
 - support for new dialogflow data format (previously api.ai)
+- language option to convert script
 
 Fixed
 -----
 - Fixed loading of default model from S3. Fixes #633
 - fixed permanent training status when training fails #652
-
+- quick fix for None "_formatter_parser" bug
 
 [0.10.1] - 2017-10-06
 ^^^^^^^^^^^^^^^^^^^^^